requirements.txt
altgraph==0.17
Appium-Python-Client==1.0.2
astor==0.8.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cpca==0.5.1
cycler==0.10.0
decorator==4.4.2
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
graphviz==0.16
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
LAC==2.1.1
matplotlib==3.3.3
nltk==3.5
numpy==1.19.4
objgraph==3.5.0
opencv-python==4.4.0.46
paddlepaddle==1.8.5
pandas==1.2.0
pathlib==1.0.1
pefile==2019.4.18
Pillow==8.0.1
prettytable==2.0.0
protobuf==3.14.0
pyahocorasick==1.4.0
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
PyYAML==5.4
qiniu==7.3.0
QtAwesome==1.0.2
QtPy==1.9.0
qtwidgets==0.17
rarfile==4.0
regex==2020.11.13
requests==2.25.0
scipy==1.6.0
selenium==3.141.0
six==1.15.0
style==1.1.0
tqdm==4.56.0
update==0.0.1
urllib3==1.26.2
wcwidth==0.2.5
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
